monitor/requirements.txt
rich>=13.0.0
watchdog>=3.0.0
pyyaml>=6.0
psutil>=5.9.0
python-dateutil>=2.8
